+++ src/protocol/Edelstein.Protocol.Gameplay.Game/Objects/NPC/StructuredNPCInfo.cs
@@ -10,7 +10,7 @@ public record StructuredNPCInfo : StructuredBasePacket
     [FieldOrder(1)] public short X { get; init; }
     [FieldOrder(2)] public short Y { get; init; }
     [FieldOrder(3)] public byte MoveAction { get; init; }
-    [FieldOrder(4)] public byte Foothold { get; init; }
+    [FieldOrder(4)] public short Foothold { get; init; }
     
     [FieldOrder(5)] public short RangeMin { get; init; }
     [FieldOrder(6)] public short RangeMax { get; init; }
